refactor(auth): fold the project auth seam into the route wrapper
