+++ web/src/features/public-api/server/createAuthedProjectAPIRoute.ts
@@ -1,12 +1,7 @@
-import crypto from "node:crypto";
 import { type NextApiRequest, type NextApiResponse } from "next";
 import { type ZodType, type z } from "zod";
-import { ApiAuthService } from "@/src/features/public-api/server/apiAuth";
-import { prisma } from "@langfuse/shared/src/db";
 import {
-  redis,
   type AuthHeaderValidVerificationResult,
-  type ApiAccessLevel,
   traceException,
   logger,
   contextWithLangfuseProps,
@@ -31,11 +26,17 @@ import {
 } from "./structuredPublicApiErrorContract";
 import { clickHouseRouteForRequest } from "@/src/features/public-api/server/clickHouseRequestTags";
 import { attachDeprecation } from "@/src/features/public-api/server/deprecations";
-import { verifyAuth as verifyAuthWithPolicy } from "@/src/features/auth/policy/shadow.projects";
+import {
+  verifyAuth as verifyLegacyAuth,
+  type RouteAccessLevel,
+} from "@/src/features/public-api/server/verifyProjectApiKeyAuth";
+import { enforceProjectAuth } from "@/src/features/auth/policy/enforcement.projects";
+import {
+  diffResults,
+  legacyFromStatus,
+  recordCoverage,
+} from "@/src/features/auth/policy/shadow";
 import { type ProjectAction } from "@/src/features/auth/policy/types";
-
-/** Access levels that can be accepted by project-scoped API routes. */
-type RouteAccessLevel = Exclude<ApiAccessLevel, "organization">;
 
 // Next's res.json uses JSON.stringify; V8 throws this when the JSON string
 // exceeds the engine limit. Keep this check scoped to the response write.
@@ -106,210 +107,78 @@ export type AuthedProjectAPIRouteConfig<
   }) => Promise<z.infer<TResponse>>;
 };
 
-/**
- * Verifies API key authentication (Basic or Bearer) using ApiAuthService.
- *
- * Delegates to ApiAuthService.verifyAuthHeaderAndReturnScope which handles
- * both Basic auth (public + secret key) and Bearer auth (public key only).
- * The caller controls which access levels are accepted via allowedAccessLevels.
- *
- * @param authHeader - The Authorization header from the request
- * @param allowedAccessLevels - Access levels to accept (default: ["project"])
- * @returns An auth scope object with the verified access level
- * @throws Error with appropriate message if authentication fails
- */
-async function verifyApiKeyAuth(
-  authHeader: string | undefined,
-  allowedAccessLevels: RouteAccessLevel[] = ["project"],
-  allowInAppAgentKey = false,
-): Promise<
-  AuthHeaderValidVerificationResult & {
-    scope: { projectId: string; accessLevel: RouteAccessLevel };
-  }
-> {
-  const regularAuth = await new ApiAuthService(
-    prisma,
-    redis,
-  ).verifyAuthHeaderAndReturnScope(authHeader, { allowInAppAgentKey });
+/** verifyAuth is the project seam: legacy decides in legacy/shadow (byte-identical), and the new PDP gates the legacy scope in enforce. */
+export async function verifyAuth(
+  params: VerifyAuthParams,
+): Promise<LegacyResult> {
+  const legacy = await runLegacyAuth(params);
 
-  if (!regularAuth.validKey) {
-    throw { status: 401, message: regularAuth.error };
+  // legacy mode skips the new pipeline entirely so self-host does no extra auth work
+  if (env.API_AUTH_MIGRATION === "legacy") {
+    if (!legacy.ok) throw legacy.error;
+    return legacy.auth;
   }
 
-  if (
-    !(allowedAccessLevels as ApiAccessLevel[]).includes(
-      regularAuth.scope.accessLevel,
-    )
-  ) {
-    throw {
-      status: 403,
-      message: "Access denied - insufficient permissions for this endpoint",
-    };
-  }
-
-  if (!regularAuth.scope.projectId) {
-    throw {
-      status: 403,
-      message:
-        "Project ID not found for API token. Are you using an organization key?",
-    };
-  }
-
-  return regularAuth as AuthHeaderValidVerificationResult & {
-    scope: { projectId: string; accessLevel: RouteAccessLevel };
-  };
-}
-
-/**
- * Verifies admin API key authentication for self-hosted instances.
- *
- * This function checks if the request contains valid admin API key credentials:
- * 1. Authorization header must be Bearer token format with ADMIN_API_KEY value
- * 2. x-langfuse-admin-api-key header must match ADMIN_API_KEY env var exactly (for redundancy)
- * 3. x-langfuse-project-id header must be present and specify a valid project ID
- * 4. NEXT_PUBLIC_LANGFUSE_CLOUD_REGION must NOT be set (self-hosted instances only)
- *
- * The ADMIN_API_KEY must be set as an environment variable on the server.
- * This authentication method is intended for administrative operations on self-hosted instances.
- *
- * @param req - The Next.js API request
- * @returns An auth scope object if successful, null if admin auth is not being attempted
- * @throws Error with appropriate status code if admin auth fails
- */
-async function verifyAdminApiKeyAuth(req: NextApiRequest): Promise<
-  | (AuthHeaderValidVerificationResult & {
-      scope: { projectId: string; accessLevel: "project" };
-    })
-  | null
-> {
-  const authHeader = req.headers.authorization;
-  const adminApiKeyHeader = req.headers["x-langfuse-admin-api-key"];
-  const projectIdHeader = req.headers["x-langfuse-project-id"];
-
-  // If not attempting admin auth, return null to proceed with regular auth
-  if (!authHeader?.startsWith("Bearer ") || !adminApiKeyHeader) return null;
-
-  // Verify this is a self-hosted instance (not Langfuse Cloud)
-  if (env.NEXT_PUBLIC_LANGFUSE_CLOUD_REGION) {
-    throw {
-      status: 403,
-      message: "Admin API key auth is not available on Langfuse Cloud",
-    };
-  }
-
-  // Verify ADMIN_API_KEY is configured
-  const adminApiKey = env.ADMIN_API_KEY;
-  if (!adminApiKey) {
-    throw {
-      status: 500,
-      message: "Admin API key is not configured on this instance",
-    };
-  }
-
-  // Extract Bearer token
-  const bearerToken = authHeader.replace("Bearer ", "");
-
-  // Verify both the Bearer token and header match the ADMIN_API_KEY.
-  // Keep this comparison in sync with the admin-key check in
-  // web/src/ee/features/admin-api/server/adminApiAuth.ts.
-  try {
-    // timingSafeEqual throws on different input lengths, handle accordingly
-    const bearerTokenEqual = crypto.timingSafeEqual(
-      Buffer.from(bearerToken),
-      Buffer.from(adminApiKey),
-    );
-    const headerEqual = crypto.timingSafeEqual(
-      Buffer.from(String(adminApiKeyHeader)),
-      Buffer.from(adminApiKey),
-    );
-    const isEqual = bearerTokenEqual && headerEqual;
-
-    if (!isEqual) throw Error();
-  } catch {
-    throw { status: 401, message: "Invalid admin API key" };
-  }
-
-  // Verify project ID header is present
-  if (!projectIdHeader || typeof projectIdHeader !== "string") {
-    throw {
-      status: 400,
-      message:
-        "x-langfuse-project-id header is required for admin API key authentication",
-    };
-  }
-
-  // Verify project exists
-  const project = await prisma.project.findUnique({
-    where: { id: projectIdHeader, deletedAt: null },
-    select: { id: true, orgId: true },
+  const authz = await enforceProjectAuth({
+    headers: params.req.headers,
+    action: params.action,
+    allowInAppAgentKey: params.allowInAppAgentKey,
+    isAdminApiKeyAuthAllowed: params.isAdminApiKeyAuthAllowed,
   });
 
-  if (!project) {
-    throw { status: 404, message: "Project not found" };
+  if (env.API_AUTH_MIGRATION === "shadow") {
+    recordCoverage(params.name);
+    diffResults(authz, legacyFromStatus(legacy.status), {
+      seam: "project_route",
+      action: params.action,
+    });
   }
-
-  // Return auth scope matching the regular auth structure
-  return {
-    validKey: true,
-    scope: {
-      projectId: project.id,
-      accessLevel: "project" as const,
-      orgId: project.orgId,
-      plan: "oss",
-      rateLimitOverrides: [],
-      apiKeyId: "ADMIN_API_KEY", // Special identifier for audit logging
-      publicKey: "ADMIN_API_KEY",
-      isIngestionSuspended: false,
-      isInAppAgentKey: false,
-    },
-  };
+  if (env.API_AUTH_MIGRATION === "enforce" && !authz.success) {
+    throw { status: authz.error.httpCode, message: authz.error.message };
+  }
+  if (!legacy.ok) throw legacy.error;
+  return legacy.auth;
 }
 
-/**
- * Verifies authentication for API routes with support for both regular API key
- * auth (Basic or Bearer) and admin API key auth.
- *
- * This is the main authentication entry point that delegates to either admin
- * or regular API key auth based on the configuration and request headers.
- *
- * @param req - The Next.js API request
- * @param isAdminApiKeyAuthAllowed - Whether to allow admin API key authentication
- * @param allowedAccessLevels - Access levels to accept for regular API key auth
- * @returns An auth scope object with the verified access level
- * @throws Error with appropriate status code if authentication fails
- */
-export async function verifyAuth(
-  req: NextApiRequest,
-  isAdminApiKeyAuthAllowed: boolean,
-  allowedAccessLevels: RouteAccessLevel[] = ["project"],
-  allowInAppAgentKey = false,
-): Promise<
-  AuthHeaderValidVerificationResult & {
-    scope: { projectId: string; accessLevel: RouteAccessLevel };
-  }
-> {
-  if (isAdminApiKeyAuthAllowed) {
-    // Try admin API key authentication first
-    const adminAuth = await verifyAdminApiKeyAuth(req);
-    if (adminAuth) {
-      // Admin auth succeeded
-      return adminAuth;
-    }
-    // Admin auth not attempted, fall back to regular API key auth
-    return await verifyApiKeyAuth(
-      req.headers.authorization,
-      allowedAccessLevels,
-      allowInAppAgentKey,
+/** runLegacyAuth runs the legacy verify and captures its throw as a value with the status it reported. */
+async function runLegacyAuth(
+  params: VerifyAuthParams,
+): Promise<LegacyDecision> {
+  try {
+    const auth = await verifyLegacyAuth(
+      params.req,
+      params.isAdminApiKeyAuthAllowed ?? false,
+      params.allowedAccessLevels ?? ["project"],
+      params.allowInAppAgentKey ?? false,
     );
+    return { ok: true, status: 200, auth };
+  } catch (error) {
+    const status = (error as { status?: unknown }).status;
+    return {
+      ok: false,
+      status: typeof status === "number" ? status : 500,
+      error,
+    };
   }
-
-  // Only regular API key auth is allowed
-  return await verifyApiKeyAuth(
-    req.headers.authorization,
-    allowedAccessLevels,
-    allowInAppAgentKey,
-  );
 }
+
+/** VerifyAuthParams is the request plus the route's action and legacy auth options. */
+export type VerifyAuthParams = {
+  req: NextApiRequest;
+  name: string;
+  action: ProjectAction;
+  isAdminApiKeyAuthAllowed?: boolean;
+  allowedAccessLevels?: RouteAccessLevel[];
+  allowInAppAgentKey?: boolean;
+};
+
+/** LegacyResult is the legacy verify's verified scope. */
+type LegacyResult = Awaited<ReturnType<typeof verifyLegacyAuth>>;
+
+/** LegacyDecision is the legacy verify captured as a value: the verified scope, or the status + error to re-throw. */
+type LegacyDecision =
+  | { ok: true; status: 200; auth: LegacyResult }
+  | { ok: false; status: number; error: unknown };
 
 export const createAuthedProjectAPIRoute = <
   TQuery extends ZodType<any>,
@@ -350,7 +219,7 @@ export const createAuthedProjectAPIRoute = <
 
     // Verify authentication (API key or admin API key)
     try {
-      auth = await verifyAuthWithPolicy({
+      auth = await verifyAuth({
         req,
         name: routeConfig.name,
         action: routeConfig.action,
